@@ -16,11 +16,11 @@
     "nodemon": "*",
     "prop-types": "^15.5.10",
     "querystring": "^0.2.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
     "react-flip-move": "^2.9.15",
     "react-player": "^0.21.0",
-    "react-router": "^2.8.1",
+    "react-router-dom": "^4.2.2",
     "react-scripts": "1.0.12",
     "screenfull": "^3.3.1",
     "spotify-web-api-node": "^2.4.0",
