@@ -1,9 +1,25 @@
 import React, { Component } from 'react'
-import { browserHistory } from 'react-router'
 
 class Redirect extends Component {
   componentWillMount = () => {
-    browserHistory.replace('/app/stream')
+    console.log(this.props)
+    console.log(localStorage.getItem('token'))
+    if (
+      !localStorage.getItem('token') ||
+      (localStorage.getItem('token') !==
+        this.props.location.hash.slice(14, -34) &&
+        this.props.location.hash)
+    ) {
+      console.log('Ny token')
+      let newHash = this.props.location.hash.slice(14, -34)
+      if (newHash) {
+        localStorage.setItem('token', newHash)
+        console.log('SUCCESS!')
+        this.props.history.push('/app/stream')
+      } else {
+        this.props.history.push('/')
+      }
+    }
   }
   render() {
     return <div />
