@@ -1,12 +1,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import {
-  Router,
+  BrowserRouter as Router,
+  Switch,
   Route,
-  hashHistory,
   IndexRoute,
-  browserHistory
-} from 'react-router'
+  history
+} from 'react-router-dom'
 
 import App from './components/App'
 import Home from './components/Home'
@@ -18,17 +18,27 @@ import './stylesheets/main.scss'
 
 const Root = () => {
   return (
-    <Router history={browserHistory}>
-      <Route>
-        <Route path="/" component={Login} />
-        <Route path="/app" component={App}>
-          <IndexRoute component={Redirect} />
-          <Route path="stream" component={Home} />
-          <Route path="search/:id" component={Search} />
-        </Route>
-      </Route>
+    <Router>
+      <Switch>
+        <Route exact path="/" component={Login} />
+        <Route exact path="/app" component={Redirect} />
+        <Route path="/app/stream" component={Parent} />
+      </Switch>
     </Router>
   )
+}
+
+const renderRightComponent = props => {
+  console.log(props)
+  if ((props.match.url = '/app/stream')) {
+    return <Home />
+  } else if ((props.match.url = 'app/stream/search')) {
+    return <Search />
+  }
+}
+
+const Parent = props => {
+  return <App {...props}>{renderRightComponent(props)}</App>
 }
 
 ReactDOM.render(<Root />, document.querySelector('#root'))
